requirements.txt
# Dépendances Python pour le projet Supply Chain Analytics

# Core Data Processing
pandas==2.1.4
numpy==1.24.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Apache Airflow
apache-airflow==2.7.3
apache-airflow-providers-postgres==5.7.1

# dbt
dbt-core==1.7.4
dbt-postgres==1.7.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Data Visualization (pour les notebooks)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Development & Testing
pytest==7.4.3
black==23.12.1
flake8==6.1.0
